SQLwrite Update 2018/5/30
Data type: float -> double

+++ SQLwrite/src/request/RechargeRequest.java
@@ -16,7 +16,7 @@ public final class RechargeRequest extends Request
 	 * @param requestTime «Î«Û ±º‰
 	 * @throws RequestException
 	 */
-	public RechargeRequest(String requestID, String userID, float amount, boolean method, String requestTime) throws RequestException
+	public RechargeRequest(String requestID, String userID, double amount, boolean method, String requestTime) throws RequestException
 	{
 		super(requestID, userID, "0", amount, false, method, requestTime);
 	}
